input: Handle PVHv2 like PV guests.
This is done because, to input_server at least, the handling will be the
same: setup/connect to vkbd frontend.
It is less than ideal too. PVonHVM has been around for a while, and
nothing forbids an HVM guest to use vkbd. With current rational, PVonHVM
are forbidden on HVM.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>