input: Refactor send_config.
-Warray-bound detects an access out of bounds in keybits (through
btnbits). This might have been introduced with the switch to 64-bit
since now the unsigned long are 8-byte instead of 4-byte.

BTN_WORDS is hard-set to 3 which does not reflect its actual size, it
should be 2 (BTN_MISC to KEY_OK).

In any case, the send_config() function is quite obscure and dependent
on more constant-macros. This is an attempt at refactoring:
- Use bitmask macros,
- Use event-code constant instead of arbitrary masks,
- Rely on eventtypes bitmask to construct the dmbus packet,
- Add comments.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1709